Prevent clients to upload any other file types except images

+++ src/main/kotlin/com/andrii_a/muze/routes/ArtistsRoute.kt
@@ -2,10 +2,7 @@ package com.andrii_a.muze.routes
 
 import com.andrii_a.muze.domain.model.ArtistResponse
 import com.andrii_a.muze.domain.repository.ArtistsRepository
-import com.andrii_a.muze.util.DEFAULT_INITIAL_PAGE
-import com.andrii_a.muze.util.DEFAULT_PAGE_SIZE
-import com.andrii_a.muze.util.Environment
-import com.andrii_a.muze.util.save
+import com.andrii_a.muze.util.*
 import io.ktor.http.*
 import io.ktor.http.content.*
 import io.ktor.server.application.*
@@ -83,6 +80,10 @@ fun Route.artistsRoute(repository: ArtistsRepository) {
                         }
 
                         is PartData.FileItem -> {
+                            if (!partData.isImage) {
+                                throw UploadedFileNotAnImageException()
+                            }
+
                             fileName = partData.save(Environment.artistsPortraitsDirectoryUrl)
                         }
 
@@ -100,10 +101,28 @@ fun Route.artistsRoute(repository: ArtistsRepository) {
                     bio = bio
                 )
 
-                call.respond(HttpStatusCode.OK, portraitImageUrl)
+                call.respond(
+                    message = portraitImageUrl,
+                    status = HttpStatusCode.OK
+                )
+
             } catch (ex: Exception) {
-                File("/$fileName").delete()
-                call.respond(HttpStatusCode.InternalServerError, "Unable to receive data. Try again later")
+                when (ex) {
+                    is UploadedFileNotAnImageException -> {
+                        call.respond(
+                            message = ex.message.orEmpty(),
+                            status = HttpStatusCode.BadRequest
+                        )
+                    }
+
+                    else -> {
+                        File("/$fileName").delete()
+                        call.respond(
+                            message = "Unable to receive data. Try again later",
+                            status = HttpStatusCode.InternalServerError
+                        )
+                    }
+                }
             }
         }
